refactor(color_picker/cli): rename files to get rid of multiple extensions
